--- app/javascript/src/components/Authentication/constants.js
@@ -1,38 +1,0 @@
-import * as yup from "yup";
-
-export const LOGIN_FORM_INITIAL_VALUES = {
-  email: "",
-  password: "",
-};
-
-export const RESET_PASSWORD_FORM_INITIAL_VALUES = {
-  email: "",
-};
-
-export const SIGNUP_FORM_INITIAL_VALUES = {
-  email: "",
-  firstName: "",
-  lastName: "",
-  password: "",
-  passwordConfirmation: "",
-};
-
-export const LOGIN_FORM_VALIDATION_SCHEMA = yup.object().shape({
-  email: yup.string().email("Invalid email address").required("Required"),
-  password: yup.string().required("Required"),
-});
-
-export const RESET_PASSWORD_FORM_VALIDATION_SCHEMA = yup.object().shape({
-  email: yup.string().email("Invalid email address").required("Required"),
-});
-
-export const SIGNUP_FORM_VALIDATION_SCHEMA = yup.object().shape({
-  email: yup.string().email("Invalid email address").required("Required"),
-  firstName: yup.string().required("Required"),
-  lastName: yup.string().required("Required"),
-  password: yup.string().required("Required"),
-  passwordConfirmation: yup
-    .string()
-    .oneOf([yup.ref("password")], "Passwords must match")
-    .required("Required"),
-});
