--- app/javascript/src/components/Dashboard/Notes/constants.js
@@ -1,26 +1,0 @@
-import * as yup from "yup";
-
-export const NOTES_FORM_INITIAL_FORM_VALUES = {
-  title: "",
-  description: "",
-};
-
-export const NOTES_FORM_VALIDATION_SCHEMA = yup.object().shape({
-  title: yup.string().required("Title is required"),
-  description: yup.string().required("Description is required"),
-});
-
-export const NOTES_TABLE_COLUMN_DATA = [
-  {
-    title: "Title",
-    dataIndex: "title",
-    key: "title",
-    width: "30%",
-  },
-  {
-    title: "Description",
-    dataIndex: "description",
-    key: "description",
-    width: "70%",
-  },
-];
