--- app/javascript/src/components/Dashboard/Settings/utils.js
@@ -1,13 +1,0 @@
-import { SETTINGS_NAVLINKS } from "./navLinks";
-
-export const getActiveNavLink = key =>
-  SETTINGS_NAVLINKS.find(navlink => key === navlink.key);
-
-export const buildProfileFormInitialValues = user => ({
-  firstName: user.first_name,
-  lastName: user.last_name,
-});
-
-export const buildEmailFormInitialValues = user => ({
-  email: user.email,
-});
