+++ app/javascript/src/components/Notes/Card.jsx
