+++ app/javascript/src/components/Notes/SidePanel.jsx
