+++ app/javascript/src/components/Notes/constants.js
