--- app/javascript/src/components/SidePanelRow.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-const SidePanelRow = ({ label, count }) => (
-  <div className="ml-6 flex flex-row justify-between px-2 py-2">
-    <div>{label}</div>
-    <div>{count}</div>
-  </div>
-);
-
-export default SidePanelRow;
