--- app/javascript/src/reducers/user.js
@@ -1,14 +1,0 @@
-const userReducer = (_, { type, payload }) => {
-  switch (type) {
-    case "SET_USER": {
-      return {
-        user: payload.user,
-      };
-    }
-    default: {
-      throw new Error(`Unhandled action type: ${type}`);
-    }
-  }
-};
-
-export default userReducer;
