+++ app/javascript/stylesheets/application.scss
@@ -13,4 +13,3 @@
 @import "components/notes";
 @import "components/sidepanel";
 @import "components/card";
-@import "components/notelisting"
