--- app/javascript/stylesheets/components/_notelisting.scss
@@ -1,7 +1,0 @@
-.add-button{
-
-  margin: 0 24px 0 12px;
-}
-.add-button:hover{
-
-}
